requirements.txt
pandas>=1.5.0
numpy>=1.23.0
scikit-learn>=1.2.0
openpyxl>=3.0.0
mysql-connector-python>=8.0.0
xlrd>=2.0.0
fastapi>=0.95.0
uvicorn>=0.22.0
gunicorn>=20.1.0
python-dotenv>=1.0.0